requirements.txt
brotlipy==0.7.0
certifi==2021.5.30
cffi @ file:///opt/concourse/worker/volumes/live/4a8f7628-1917-402a-5a6a-cdafcc3ec547/volume/cffi_1625814707380/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
cryptography @ file:///opt/concourse/worker/volumes/live/fab3474b-40c6-4001-48cb-802be615e43d/volume/cryptography_1633520388016/work
greenlet==1.1.2
idna @ file:///tmp/build/80754af9/idna_1622654382723/work
importlib-metadata==4.8.1
loguru==0.5.3
numpy==1.21.2
pandas==1.3.3
powerline-status==2.7
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
PyExecJS==1.5.1
PyMySQL==1.0.2
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
PySocks @ file:///opt/concourse/worker/volumes/live/ef943889-94fc-4539-798d-461c60b77804/volume/pysocks_1605305801690/work
python-dateutil==2.8.2
pytz==2021.3
requests @ file:///tmp/build/80754af9/requests_1629994808627/work
six @ file:///tmp/build/80754af9/six_1623709665295/work
SQLAlchemy==1.4.25
typing-extensions==3.10.0.2
urllib3==1.26.7
zipp==3.6.0
